Rename: 클래스 이름을 직관적으로 수정
- CarGameManager 는 입, 출력 없이 게임 로직만 수행하므로 domain 으로 이동
